Add timing + quick_model option
- Also moved get_atom_indices, write_pdbs_of_clusters into utils.py
